Fix broken integration:show command and add format option
